requirements.txt
python-chess
matplotlib
numpy
pandas
tqdm
typing_extensions
graphviz
pygame
pygame-popup
tensorflow-gpu